fix(macos): satisfy cookie sync lint gates
